Fix for paths that may not exist.

+++ trunk/show.php
@@ -20,11 +20,25 @@ if ($_GET['state']) {
 
 $state = $_SESSION['show']['state'];
 
+// Create the episode dir if it doesn't exist
+if (!is_dir($epDir) && !mkdir($epDir, 0775)) {
+    custom_error('Error creating '.$epDir.': Please check permissions on the data directory.');
+    exit;
+}
+
+// Create the shows dir if it doesn't exist
+if (!is_dir($showDir) && !mkdir($showDir, 0775)) {
+    custom_error('Error creating '.$showDir.': Please check permissions on the data directory.');
+    exit;
+}
+
 // If a shows.txt file doesn't exist or you select update a new list of
 // shows will be grabbed from tvrage.com
 if (!file_exists($showsTxt) || $state == "update") {
     exec("modules/episode/utils/grabshowsall.pl $showsTxt $showsCountry");
     unset($_SESSION['show']['state']);
+    $_SESSION['show']['state'] = "current";
+    $state = $_SESSION['show']['state'];
     if (file_exists($showsDat))
         unlink($showsDat);
 }
